fix: Playlist JSON 수정 (#264)
