+++ src/constants/json/genres/movieAndGenre.json
@@ -5,39 +5,43 @@
   },
   {
     "movie_id": 1,
-    "genre_id": 99
+    "genre_id": 10749
   },
   {
     "movie_id": 2,
-    "genre_id": 99
+    "genre_id": 28
   },
   {
     "movie_id": 2,
-    "genre_id": 99
+    "genre_id": 14
   },
   {
     "movie_id": 3,
-    "genre_id": 99
+    "genre_id": 28
   },
   {
     "movie_id": 3,
-    "genre_id": 99
+    "genre_id": 18
   },
   {
     "movie_id": 4,
-    "genre_id": 99
+    "genre_id": 37
   },
   {
     "movie_id": 4,
-    "genre_id": 99
+    "genre_id": 18
   },
   {
     "movie_id": 4,
-    "genre_id": 99
+    "genre_id": 10751
   },
   {
     "movie_id": 5,
-    "genre_id": 99
+    "genre_id": 12
+  },
+  {
+    "movie_id": 6,
+    "genre_id": 80
   },
   {
     "movie_id": 6,
@@ -45,135 +49,131 @@
   },
   {
     "movie_id": 6,
-    "genre_id": 99
-  },
-  {
-    "movie_id": 6,
-    "genre_id": 99
+    "genre_id": 36
   },
   {
     "movie_id": 7,
-    "genre_id": 99
+    "genre_id": 9648
   },
   {
     "movie_id": 7,
-    "genre_id": 99
+    "genre_id": 27
   },
   {
     "movie_id": 7,
-    "genre_id": 99
+    "genre_id": 10770
   },
   {
     "movie_id": 8,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 9,
-    "genre_id": 99
+    "genre_id": 16
   },
   {
     "movie_id": 10,
-    "genre_id": 99
+    "genre_id": 37
   },
   {
     "movie_id": 11,
-    "genre_id": 99
+    "genre_id": 10749
   },
   {
     "movie_id": 12,
-    "genre_id": 99
+    "genre_id": 18
   },
   {
     "movie_id": 12,
-    "genre_id": 99
+    "genre_id": 14
   },
   {
     "movie_id": 13,
-    "genre_id": 99
+    "genre_id": 14
   },
   {
     "movie_id": 13,
-    "genre_id": 99
+    "genre_id": 27
   },
   {
     "movie_id": 14,
-    "genre_id": 99
+    "genre_id": 10752
   },
   {
     "movie_id": 14,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 14,
-    "genre_id": 99
+    "genre_id": 10402
   },
   {
     "movie_id": 15,
-    "genre_id": 99
+    "genre_id": 10752
   },
   {
     "movie_id": 15,
-    "genre_id": 99
+    "genre_id": 27
   },
   {
     "movie_id": 16,
-    "genre_id": 99
+    "genre_id": 37
   },
   {
     "movie_id": 16,
-    "genre_id": 99
+    "genre_id": 14
   },
   {
     "movie_id": 17,
-    "genre_id": 99
+    "genre_id": 16
   },
   {
     "movie_id": 18,
-    "genre_id": 99
+    "genre_id": 36
   },
   {
     "movie_id": 18,
-    "genre_id": 99
+    "genre_id": 35
   },
   {
     "movie_id": 18,
-    "genre_id": 99
+    "genre_id": 10749
   },
   {
     "movie_id": 19,
-    "genre_id": 99
+    "genre_id": 37
   },
   {
     "movie_id": 19,
-    "genre_id": 99
+    "genre_id": 14
   },
   {
     "movie_id": 19,
-    "genre_id": 99
+    "genre_id": 18
   },
   {
     "movie_id": 20,
-    "genre_id": 99
+    "genre_id": 18
   },
   {
     "movie_id": 21,
-    "genre_id": 99
+    "genre_id": 10770
   },
   {
     "movie_id": 22,
-    "genre_id": 99
+    "genre_id": 80
   },
   {
     "movie_id": 22,
-    "genre_id": 99
+    "genre_id": 12
   },
   {
     "movie_id": 23,
-    "genre_id": 99
+    "genre_id": 53
   },
   {
     "movie_id": 23,
-    "genre_id": 99
+    "genre_id": 10770
   },
   {
     "movie_id": 24,
@@ -181,7 +181,11 @@
   },
   {
     "movie_id": 24,
-    "genre_id": 99
+    "genre_id": 28
+  },
+  {
+    "movie_id": 25,
+    "genre_id": 10402
   },
   {
     "movie_id": 25,
@@ -189,27 +193,23 @@
   },
   {
     "movie_id": 25,
-    "genre_id": 99
-  },
-  {
-    "movie_id": 25,
-    "genre_id": 99
+    "genre_id": 10749
   },
   {
     "movie_id": 26,
-    "genre_id": 99
+    "genre_id": 35
   },
   {
     "movie_id": 26,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 27,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 28,
-    "genre_id": 99
+    "genre_id": 36
   },
   {
     "movie_id": 28,
@@ -217,7 +217,7 @@
   },
   {
     "movie_id": 29,
-    "genre_id": 99
+    "genre_id": 10751
   },
   {
     "movie_id": 29,
@@ -225,94 +225,94 @@
   },
   {
     "movie_id": 30,
-    "genre_id": 99
+    "genre_id": 10770
   },
   {
     "movie_id": 30,
-    "genre_id": 99
+    "genre_id": 27
   },
   {
     "movie_id": 31,
-    "genre_id": 99
+    "genre_id": 37
   },
   {
     "movie_id": 31,
-    "genre_id": 99
+    "genre_id": 16
   },
   {
     "movie_id": 31,
-    "genre_id": 99
+    "genre_id": 35
   },
   {
     "movie_id": 32,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 32,
-    "genre_id": 99
+    "genre_id": 35
   },
   {
     "movie_id": 32,
-    "genre_id": 99
+    "genre_id": 37
   },
   {
     "movie_id": 33,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 33,
-    "genre_id": 99
+    "genre_id": 35
   },
   {
     "movie_id": 34,
-    "genre_id": 99
+    "genre_id": 10751
   },
   {
     "movie_id": 35,
-    "genre_id": 99
+    "genre_id": 10751
   },
   {
     "movie_id": 35,
-    "genre_id": 99
+    "genre_id": 35
   },
   {
     "movie_id": 36,
-    "genre_id": 99
+    "genre_id": 36
   },
   {
     "movie_id": 36,
-    "genre_id": 99
+    "genre_id": 10402
   },
   {
     "movie_id": 37,
-    "genre_id": 99
+    "genre_id": 10752
   },
   {
     "movie_id": 37,
-    "genre_id": 99
+    "genre_id": 27
   },
   {
     "movie_id": 37,
-    "genre_id": 99
+    "genre_id": 878
   },
   {
     "movie_id": 38,
-    "genre_id": 99
+    "genre_id": 10749
   },
   {
     "movie_id": 39,
-    "genre_id": 99
+    "genre_id": 10751
   },
   {
     "movie_id": 40,
-    "genre_id": 99
+    "genre_id": 18
   },
   {
     "movie_id": 40,
-    "genre_id": 99
+    "genre_id": 80
   },
   {
     "movie_id": 40,
-    "genre_id": 99
+    "genre_id": 36
   }
 ]
